+++ astarte-dashboard/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add pagination to the device list page.
+
 ### Changed
 - Layout and palette restyle.
 - Configuration accepts both base Astarte API URL and specific component URLs.